Skip unaligned access performance check
See https://lore.kernel.org/lkml/20230801221138.2086734-2-evan@rivosinc.com/
for details about how this works.

Most importantly it mentions that until the kernel gets the ability to
be delegated unaligned access traps instead of the "firmware silently
handling it," we can't detect if unaligned accesses are unsupported.

Until such a time therefore that ogsbi and nommulinux either begin
to emulate unaligned accesses, or support a new future SBI feature
to delegate unaligned access traps, we are stuck.

So avoid the issue entirely by hardcoding it to _UNSUPPORTED, removing
the code responsible for the check altogether.